+++ src/components/options/columnstyles/ColumnStylesEditor.tsx
@@ -20,7 +20,7 @@ import {
 import { ColumnStyleItem } from './ColumnStyleItem';
 import { Threshold } from '../thresholds/types';
 import { getColumnHints } from './columnHints';
-import { ColumnStyleColoring, DateFormats } from 'types';
+import { ColumnAlignment, ColumnStyleColoring, DateFormats } from 'types';
 import { TrackerAdapter, useTracker } from 'hooks/useTracker';
 
 const columnStyleAdapter: TrackerAdapter<ColumnStyleItemTracker, ColumnStyleItemType> = {
@@ -87,6 +87,7 @@ export const ColumnStylesEditor: React.FC<StandardEditorProps<ColumnStyleItemTyp
       nameOrRegex: '',
       order,
       enabled: true,
+      align: ColumnAlignment.DEFAULT,
       dateStyle: { dateFormat: DateFormats[0].value } as ColumnStyleDate,
       hiddenStyle: {} as ColumnStyleHidden,
       metricStyle: {